requirements.txt
streamlit==1.24.1
requests_oauthlib==2.0.0
google-auth==2.35.0
google-auth-oauthlib==0.4.6
Pillow  # Add this line if it's needed
